full precision float changes for coordinates

+++ openfold3/model_implementations/af3_all_atom/runner.py
@@ -7,9 +7,9 @@ from openfold3.core.metrics.validation_all_atom import (
     get_validation_metrics,
 )
 from openfold3.core.runners.model_runner import ModelRunner
+from openfold3.core.utils.atomize_utils import broadcast_token_feat_to_atoms
 from openfold3.core.utils.lr_schedulers import AlphaFoldLRScheduler
 from openfold3.core.utils.tensor_utils import tensor_tree_map
-from openfold3.core.utils.atomize_utils import broadcast_token_feat_to_atoms
 from openfold3.model_implementations.af3_all_atom.config.base_config import config
 from openfold3.model_implementations.af3_all_atom.model import AlphaFold3
 from openfold3.model_implementations.registry import register_model
@@ -27,8 +27,8 @@ class AlphaFold3AllAtom(ModelRunner):
     ):
         metrics = {}
 
-        gt_coords = batch["ground_truth"]["atom_positions"] 
-        pred_coords = outputs["x_pred"]
+        gt_coords = batch["ground_truth"]["atom_positions"].float()
+        pred_coords = outputs["x_pred"].float()
         all_atom_mask = batch["ref_mask"]
         token_mask = batch["token_mask"]
         num_atoms_per_token = batch["num_atoms_per_token"]
